requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
fusion-solar-py==0.0.9
idna==3.10
mock==5.1.0
python-dotenv==1.0.1
requests==2.32.3
RPi.GPIO==0.7.1
urllib3==2.3.0
